Module linefeed::chars

Provides utilities for manipulating character values

Constants

DELETE

Character value generated by the Backspace key

EOF

Character value indicating end-of-file

ESCAPE

Character value generated by the Escape key

RUBOUT

Character value generated by the Backspace key on some systems

Functions

ctrl

Returns a control character for the given character.

escape_sequence

Returns a character sequence escaped for user-facing display.

is_ctrl

Returns whether the given character is a control character.

is_printable

Returns whether the character is printable.

meta

Returns a meta sequence for the given character.

parse_char_name

Returns a character name as a key sequence, e.g. Control-x or Meta-x.

unctrl

Returns the printable character corresponding to the given control character.

unctrl_lower

Returns the lowercase character corresponding to the given control character.