Trait linefeed::complete::Completer

pub trait Completer<Term: Terminal>: Send + Sync {
    fn complete(
        &self, 
        word: &str, 
        prompter: &Prompter<'_, '_, Term>, 
        start: usize, 
        end: usize
    ) -> Option<Vec<Completion>>;

    fn word_start(
        &self, 
        line: &str, 
        end: usize, 
        prompter: &Prompter<'_, '_, Term>
    ) -> usize { ... }

    fn quote<'a>(&self, word: &'a str) -> Cow<'a, str> { ... }

    fn unquote<'a>(&self, word: &'a str) -> Cow<'a, str> { ... }
}

Performs completion for Prompter when triggered by a user input sequence

Required methods

fn complete(
&self,
word: &str,
prompter: &Prompter<'_, '_, Term>,
start: usize,
end: usize
) -> Option<Vec<Completion>>

Returns the set of possible completions for the prefix word.

Provided methods

fn word_start(
&self,
line: &str,
end: usize,
prompter: &Prompter<'_, '_, Term>
) -> usize

Returns the starting position of the word under the cursor.

The default implementation uses Prompter::word_break_chars() to detect the start of a word.

fn quote<'a>(&self, word: &'a str) -> Cow<'a, str>

Quotes a possible completion for insertion into input.

The default implementation returns the word, as is.

fn unquote<'a>(&self, word: &'a str) -> Cow<'a, str>

Unquotes a piece of user input before searching for completions.

The default implementation returns the word, as is.