Struct linefeed::memory::Lines

pub struct Lines<'a> { /* fields omitted */ }

Iterator over lines in a MemoryTerminal buffer.

Note that while this value behaves as an iterator, it cannot implement the Iterator trait because its yielded values borrow self.

Methods

impl<'a> Lines<'a>

fn next(&mut self) -> Option<&[char]>

Returns the next line in the buffer.

fn lines_remaining(&self) -> usize

Returns the number of lines remaining in the iterator.