Struct linefeed::memory::MemoryTerminal

pub struct MemoryTerminal { /* fields omitted */ }

Implements an in-memory Terminal interface

The contents of a MemoryTerminal are shared. That is, cloning a MemoryTerminal value will share the contained terminal buffer.

Implementations

impl MemoryTerminal

pub fn new() -> MemoryTerminal

Returns a new MemoryTerminal with the default buffer size.

pub fn with_size(size: Size) -> MemoryTerminal

Returns a new MemoryTerminal with the given buffer size.

Panics

If either of the lines or columns fields are 0.

pub fn clear_all(&self)

Clears the terminal buffer and places the cursor at (0, 0).

pub fn clear_to_end(&self)

Clears all characters beginning at the cursor and ending at buffer end.

pub fn clear_input(&self)

Clears the input buffer.

pub fn has_input(&self) -> bool

Returns whether any input remains to be read.

pub fn lines(&self) -> Lines<'_>

Returns an iterator over lines in the buffer.

Notes

The returned iterator immutably borrows the contents of the MemoryTerminal. Attempting to perform a mutating operation on the parent MemoryTerminal while the Lines iterator lives will cause a panic.

pub fn move_up(&self, n: usize)

Moves the cursor up n cells.

pub fn move_down(&self, n: usize)

Moves the cursor down n cells.

pub fn move_left(&self, n: usize)

Moves the cursor left n cells.

pub fn move_right(&self, n: usize)

Moves the cursor right n cells.

pub fn move_to_first_column(&self)

Moves the cursor to the first column of the current line.

pub fn push_input(&self, s: &str)

Pushes a character sequence to the back of the input queue.

pub fn read_input(&self, buf: &mut [u8]) -> usize

Reads some input from the input buffer.

pub fn resize(&self, new_size: Size)

Changes the size of the terminal buffer. The buffer will be truncated or filled with spaces, as necessary.

Panics

If either of the lines or columns fields are 0 or if the area exceeds usize maximum.

pub fn scroll_up(&self, n: usize)

Moves the contents of the buffer up n lines. The first n lines of text will be erased.

pub fn cursor(&self) -> (usize, usize)

Returns the (line, column) position of the cursor.

pub fn set_cursor_mode(&self, mode: CursorMode)

Sets the cursor mode.

pub fn cursor_mode(&self) -> CursorMode

Returns the cursor mode.

pub fn size(&self) -> Size

Returns the size of the terminal buffer.

pub fn write(&self, s: &str)

Writes some text into the buffer.

If the text extends beyond the length of the current line without a newline character ('\n'), the extraneous text will be dropped.