Module linefeed::reader

Provides high-level line editing interface

Structs

BindingIter
Iterator over Reader bindings

HistoryIter
Iterator over Reader history entries

LogSender
A handle that allows other threads to send messages to a Reader without causing display errors in a currently open prompt.

Reader
Interactively reads user input

VariableIter
Iterator over Reader variable values

Enums

ReadResult
The result of Reader::read_line

Variable
Represents a Reader variable of a given type

Constants

BLINK_TIMEOUT_MS
Timeout, in milliseconds, to wait for input when "blinking"

END_INVISIBLE
Indicates the end of a series of invisible characters in the prompt

KEYSEQ_TIMEOUT_MS
Default keyseq_timeout, in milliseconds

MAX_HISTORY
Default maximum size of history

POLL_LOG_INTERVAL_MS
Default intervals at which the Reader will check for log messages.

START_INVISIBLE
Indicates the start of a series of invisible characters in the prompt

STRING_CHARS
Default set of string characters

WORD_BREAK_CHARS
Default set of word break characters