Struct linefeed::terminal::DefaultTerminal

pub struct DefaultTerminal(_);

Default Terminal interface

Implementations

impl DefaultTerminal

pub fn new() -> Result<DefaultTerminal>

Opens access to the terminal device associated with standard output.

pub fn stderr() -> Result<DefaultTerminal>

Opens access to the terminal device associated with standard error.