Enum linefeed::command::Command

pub enum Command {

    Abort,
    AcceptLine,
    Complete,
    InsertCompletions,
    PossibleCompletions,
    MenuComplete,
    MenuCompleteBackward,
    DigitArgument,
    SelfInsert,
    TabInsert,
    OverwriteMode,
    InsertComment,
    BackwardChar,
    ForwardChar,
    CharacterSearch,
    CharacterSearchBackward,
    BackwardWord,
    ForwardWord,
    BackwardKillLine,
    KillLine,
    BackwardKillWord,
    KillWord,
    UnixWordRubout,
    ClearScreen,
    BeginningOfLine,
    EndOfLine,
    BackwardDeleteChar,
    DeleteChar,
    TransposeChars,
    TransposeWords,
    BeginningOfHistory,
    EndOfHistory,
    NextHistory,
    PreviousHistory,
    ForwardSearchHistory,
    ReverseSearchHistory,
    HistorySearchForward,
    HistorySearchBackward,
    QuotedInsert,
    Yank,
    YankPop,
    Custom(Cow<'static, str>),
    Macro(Cow<'static, str>),

}

Represents a command to modify Reader state

Variants

Abort

Abort history search

AcceptLine

Accepts the current input line

Complete

Perform completion

InsertCompletions

Insert all completions into the input buffer

PossibleCompletions

Show possible completions

MenuComplete

Insert the next possible completion

MenuCompleteBackward

Insert the previous possible completion

DigitArgument

Begin numeric argument input

SelfInsert

Insert character or sequence at the cursor

TabInsert

Inserts a tab character

OverwriteMode

Toggles insert/overwrite mode

InsertComment

Insert a comment and accept input

BackwardChar

Move the cursor backward one character

ForwardChar

Move the cursor forward one character

CharacterSearch

Search for a given character

CharacterSearchBackward

Search backward for a given character

BackwardWord

Move the cursor backward one word

ForwardWord

Move the cursor forward one word

BackwardKillLine

Kill all characters before the cursor

KillLine

Kill all characters after the cursor

BackwardKillWord

Kill a word before the cursor

KillWord

Kill a word after the cursor

UnixWordRubout

Kill a word before the cursor, delimited by whitespace

ClearScreen

Clear the screen

BeginningOfLine

Move the cursor to the beginning of the line

EndOfLine

Move the cursor to the end of the line

BackwardDeleteChar

Delete one character before the cursor

DeleteChar

Delete one character after the cursor

TransposeChars

Drag the character before the cursor forward

TransposeWords

Drag the word before the cursor forward

BeginningOfHistory

Move to the first line of history

EndOfHistory

Move to the last line of history

NextHistory

Select next line in history

PreviousHistory

Select previous line in history

ForwardSearchHistory

Incremental search in history

ReverseSearchHistory

Incremental reverse search in history

HistorySearchForward

Non-incremental forward history search using input up to the cursor

HistorySearchBackward

Non-incremental backward history search using input up to the cursor

QuotedInsert

Insert literal character

Yank

Insert text into buffer from the kill ring

YankPop

Rotate the kill ring and yank the new top

Custom(Cow<'static, str>)

Custom application-defined command

Tuple Fields of Custom

0: Cow<'static, str>

Macro(Cow<'static, str>)

Execute a given key sequence

Tuple Fields of Macro

0: Cow<'static, str>

Implementations

impl Command

pub fn from_str(name: &'static str) -> Command

Constructs a command from a 'static str reference.

If the string does not refer to a built-in command, a value of Command::Custom(Borrowed(name)) will be returned.

pub fn from_string<T>(name: T) -> Command where
T: AsRef<str> + Into<String>,

Constructs a command from a non-'static string-like type.

If the string does not refer to a built-in command, a value of Command::Custom(Owned(name.into())) will be returned.

impl Command

pub fn category(&self) -> Category

Returns the category of the command