Struct linefeed::complete::Completion

pub struct Completion {
    pub completion: String,
    pub display: Option<String>,
    pub suffix: Suffix,
}

Represents a single possible completion

Fields
completion: String

Whole completion text

display: Option<String>

Listing display string; None if matches completion

suffix: Suffix

Completion suffix; replaces append character

Methods

impl Completion

pub fn simple(s: String) -> Completion

Returns a simple Completion value, with display string matching completion and using the default completion suffix.

pub fn completion(&self, def_suffix: Option<char>) -> Cow<str>

Returns the full completion string, including suffix, using the given default suffix if one is not assigned to this completion.

pub fn display(&self) -> Cow<str>

Returns the display string, including suffix

pub fn display_chars(&self) -> usize

Returns the number of characters displayed